src/test/kotlin/com/colinodell/advent2021/Day25Test.kt
package com.colinodell.advent2021

import com.colinodell.advent2021.Resources.resourceAsListOfString
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test

@DisplayName("Day 25")
class Day25Test {
    private val sample = """
        v...>>.vv>
        .vv>>.vv..
        >>.>v>...v
        >>v>>.>.v.
        v>v.vv.v..
        >.>>..v...
        .vv..>.>v.
        v.v..>>v.v
        ....v..v.>
    """.trimIndent().trim().split("\n")

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches Sample Data`() {
            val answer = Day25(sample).solvePart1()
            assertThat(answer).isEqualTo(58)
        }

        @Test
        fun `Matches Actual Data`() {
            val answer = Day25(resourceAsListOfString("day25.txt")).solvePart1()
            assertThat(answer).isEqualTo(520)
        }
    }
}